12th-College Students  The 12th-College participants will be learning the scientific and engineering practice of participation and crowd reactions. College participants will judge the experience of their performance. Participants will be judged on specific criteria. Good posture is important to balance and center the body’s weight over the feet. This can reduce the likelihood of injury and improve overall health and appearance of the performance. The flexibility enhances the athleticism of the performance and increases the range of motion in joints. Regular stretching exercises can help maintain or improve flexibility. Power in the body is essential because it provides the energy necessary for basic bodily functions, maintaining homeostasis, and enabling physical activity. The body converts four calories into energy through metabolic processes, primarily using glucose, proteins, and fats to generate adenosine triphosphate (ATP), which cells use to perform vital tasks such as muscle contraction, nerve signaling, and cellular repair.
12th-College Students

The 12th-College participants will be learning the scientific and engineering practice of participation and crowd reactions. College participants will judge the experience of their performance. Participants will be judged on specific criteria. Good posture is important to balance and center the body’s weight over the feet. This can reduce the likelihood of injury and improve overall health and appearance of the performance. The flexibility enhances the athleticism of the performance and increases the range of motion in joints. Regular stretching exercises can help maintain or improve flexibility. Power in the body is essential because it provides the energy necessary for basic bodily functions, maintaining homeostasis, and enabling physical activity. The body converts four calories into energy through metabolic processes, primarily using glucose, proteins, and fats to generate adenosine triphosphate (ATP), which cells use to perform vital tasks such as muscle contraction, nerve signaling, and cellular repair.