Grace  Participant move to the next stage of grace. Keep your back and entire body straight. This helps to distribute weight evenly, making it more comfortable. Walk heel to toe smoothly to transfer the impact to the legs instead of the ball of the foot, and relax your hips and knees to be more fluid and graceful. The participant’s spiritual healing can have a positive impact on one’s well-being. The participant can improve comfort during difficult times and foster a sense of belonging in performance routines. The participant  must have artistic ability and creativity to express ideas through movement and create new and interesting dance routines. The participant must showcase the ability to perform without touching the wood floor with their knee.
Grace

Participant move to the next stage of grace. Keep your back and entire body straight. This helps to distribute weight evenly, making it more comfortable. Walk heel to toe smoothly to transfer the impact to the legs instead of the ball of the foot, and relax your hips and knees to be more fluid and graceful. The participant’s spiritual healing can have a positive impact on one’s well-being. The participant can improve comfort during difficult times and foster a sense of belonging in performance routines. The participant  must have artistic ability and creativity to express ideas through movement and create new and interesting dance routines. The participant must showcase the ability to perform without touching the wood floor with their knee.