Judges  Judges will respond with a penalty towards your performance. The speed of the participant’s performance through space is also judged. Newton’s third law of motion states that when two bodies interact, they apply forces to one another that are equal in magnitude and opposite in direction. Body language plays an important role in communicating as it can amplify, modify, confirm, or subvert verbal expression. According to a study by psychologist Albert Mehrabian on conveying emotions and attitude, nonverbal communication could account for up to 93% of the overall message, with words contributing only 7%. If the crowd reacts negatively or critically to your performance, keep the conversation productive by asking the judge, “What’s the one thing I can do to improve moving forward?” And request coaching to help you meet the goal. Demonstrating a grateful and forward-thinking attitude in the face of critiques can turn the situation into a favorable one.
Judges

Judges will respond with a penalty towards your performance. The speed of the participant’s performance through space is also judged. Newton’s third law of motion states that when two bodies interact, they apply forces to one another that are equal in magnitude and opposite in direction. Body language plays an important role in communicating as it can amplify, modify, confirm, or subvert verbal expression. According to a study by psychologist Albert Mehrabian on conveying emotions and attitude, nonverbal communication could account for up to 93% of the overall message, with words contributing only 7%. If the crowd reacts negatively or critically to your performance, keep the conversation productive by asking the judge, “What’s the one thing I can do to improve moving forward?” And request coaching to help you meet the goal. Demonstrating a grateful and forward-thinking attitude in the face of critiques can turn the situation into a favorable one.